fixed song long changes

+++ pdmapi/src/main/java/com/example/pdmapi/Service/AlbumService.java
@@ -119,7 +119,7 @@ public class AlbumService {
                 song.setSongId(rs.getLong("song_id"));
                 song.setTitle(rs.getString("title"));
                 song.setReleaseDate(rs.getDate("release_date"));
-                song.setRuntime(rs.getInt("runtime"));
+                song.setRuntime(rs.getLong("runtime"));
                 songs.add(song);
             }
         }  catch (SQLException e) {
@@ -146,7 +146,7 @@ public class AlbumService {
                 song.setSongId(rs.getLong("song_id"));
                 song.setTitle(rs.getString("title"));
                 song.setReleaseDate(rs.getDate("release_date"));
-                song.setRuntime(rs.getInt("runtime"));
+                song.setRuntime(rs.getLong("runtime"));
             }
             return song;
         }  catch (SQLException e) {
